adjusted flashlight effect, added fading (using double now instead of int)

+++ bonus/helper_bonus.c
@@ -50,7 +50,7 @@ void	fill_image_buffer(t_image image,
 {
 	t_data	*data;
 	char	*pixel_index;
-	int		darkness_level;
+	double	darkness_level;
 
 	if (x >= 0 && y >= 0 && x < image.width && y < image.height)
 	{
@@ -65,9 +65,9 @@ void	fill_image_buffer(t_image image,
 				darkness_level = 35;
 			else if (data->apply_darkness == 2)
 				darkness_level = data->flashlight[y][x];
-			color = ((((color >> 16) & 0xFF) / darkness_level) << 16)
-				| ((((color >> 8) & 0xFF) / darkness_level) << 8)
-				| ((color & 0xFF) / darkness_level);
+			color = ((int)(((color >> 16) & 0xFF) / darkness_level) << 16)
+				| ((int)(((color >> 8) & 0xFF) / darkness_level) << 8)
+				| (int)((color & 0xFF) / darkness_level);
 		}
 		*(int *)pixel_index = color;
 	}
